fix(bypass-uniffi): expose status as i32 so Java can call getStatus()
Third Android iteration: Kotlin now compiles, but javac fails at
BypassPlugin.java:55:
    result.put("status", resp.getStatus());
                          ^ cannot find symbol

`BypassResponse.status` was `u16`. UniFFI maps u16 → Kotlin UShort,
which is declared `@JvmInline value class` — its getters are
name-mangled on the JVM (`getStatus-Mh2AYeg()` and similar) so Java
callers cannot resolve a plain `getStatus()`. Same trap applies to
any unsigned (u8/u16/u32/u64): unusable from Java.

Widen the field to i32 in the FFI surface so the Kotlin binding
emits a plain `val status: Int` and the JVM facade exposes
`int getStatus()`. HTTP status fits trivially; `i32::from(resp.status)`
keeps the conversion lossless from bypass-core's u16.

Unaffected consumers:
- src-tauri/src/lib.rs talks to bypass-core directly (still u16).
- ios/App/App/BypassPlugin.swift assigns `resp.status` into
  `[String: Any]` — Int32 vs UInt16 both bridge fine.

+++ crates/bypass-uniffi/src/lib.rs
@@ -42,9 +42,14 @@ impl From<CoreBypassError> for BypassError {
     }
 }
 
+// `status` is intentionally i32, not u16. UniFFI maps u16 → Kotlin UShort,
+// which is `@JvmInline value class` and emits name-mangled getters
+// (`getStatus-Mh2AYeg()` etc.) that Java callers cannot resolve.
+// BypassPlugin.java is Java, so we expose a signed Int that yields a
+// plain `int getStatus()` on the JVM. HTTP status fits in i32 trivially.
 #[derive(Debug, Clone, uniffi::Record)]
 pub struct BypassResponse {
-    pub status: u16,
+    pub status: i32,
     pub headers: HashMap<String, String>,
     pub body: Vec<u8>,
 }
@@ -69,7 +74,8 @@ pub fn bypass_fetch(
         let resp = client.fetch(&url, headers).await?;
 
         Ok(BypassResponse {
-            status: resp.status,
+            // bypass-core's HTTP status is u16; widen to i32 for FFI.
+            status: i32::from(resp.status),
             headers: resp.headers,
             body: resp.body,
         })
